requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
mcp>=1.2.0
playwright>=1.47.0
pytesseract>=0.3.10
Pillow>=10.4.0
reportlab>=4.2.2
python-dotenv>=1.0.1
httpx>=0.27.2
openai>=1.40.0
click>=8.1.7
